Update amount type to Double and balance to Numeric
The amount and balance types from WalletController, WalletsDAO and related SQL files have been updated from 'Int' to 'Double' for handling fractional amounts. This helps in increasing precision for wallet and transaction operations and allows operations with fractional amounts.

+++ migrations/scripts/sales_oreders.sql
@@ -3,7 +3,7 @@ CREATE TABLE sales_oreders(
     user_id int,
     name varchar(255),
     description text,
-    price int,
+    price NUMERIC(15, 2),
     category int,
     photo bytea,
     FOREIGN KEY (user_id) REFERENCES users(user_id)
