+++ src/main/kotlin/com/collect/invest/dao/entity/WalletsEntity.kt
@@ -3,6 +3,6 @@ package com.collect.invest.dao.entity
 data class WalletsEntity (
     val id: Long,
     val userId: Long,
-    val balance: Int,
+    val balance: Double,
     val status: String,
 )
